Move bulk of gen_packets main back into the direwolf package
Keeping cgo out of the cmds avoids namespace issues
